moves resources + schemas into app
