+++ backend/app/main.py
@@ -4,8 +4,8 @@ import os
 from flask import Flask, jsonify
 from flask_cors import CORS
 from flask_smorest import Api
-from resources.list import blp as ListBlueprint
-from resources.task import blp as TaskBlueprint
+from backend.app.resources.list import blp as ListBlueprint
+from backend.app.resources.task import blp as TaskBlueprint
 
 def create_app(db_url=None):
     app = Flask(__name__)
