+++ backend/app/resources/list.py
@@ -3,8 +3,8 @@ from uuid import uuid4 as UUID4
 from flask.views import MethodView
 from flask_smorest import Blueprint
 
-from app import db
-from schemas.list import ListSchema
+from backend.app import db
+from backend.app.schemas.list import ListSchema
 
 blp = Blueprint('List', 'list', url_prefix='/list', description='List operations')
